https://github.com/DM-Organization130/weather-forecast-service/issues/5
user types created, API Key controls done, usage of service functions restricted

+++ src/main/java/com/example/weatherforecastservice/controller/QueryOptionController.java
@@ -1,7 +1,9 @@
 package com.example.weatherforecastservice.controller;
 
 import com.example.weatherforecastservice.model.QueryOption;
+import com.example.weatherforecastservice.model.ServiceUser;
 import com.example.weatherforecastservice.repository.QueryOptionRepository;
+import com.example.weatherforecastservice.repository.ServiceUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -13,11 +15,16 @@ import java.util.List;
 public class QueryOptionController {
     @Autowired
     private QueryOptionRepository queryOptionRepository;
-    //repositoryi doldur boşal faksiyonlarını yaz test et ikisini de
+    @Autowired
+    private ServiceUserRepository serviceUserRepository;
 
     @GetMapping("/initqueryoption")
-    public List<QueryOption> initQueryOption()
+    public List<QueryOption> initQueryOption(String userKey)
     {
+        if (serviceUserRepository.ısThereServiceUser(userKey, (byte) 2, ServiceUser.adminKey))
+        {
+            return null;
+        }
         List<String> queryDescriptions = Arrays.asList("Measure", "Day", "ThreeDay", "Weekly", "Monthly", "CityQuery");
         if (queryOptionRepository.count() == 0) {
             for (String s : queryDescriptions) {
@@ -30,8 +37,12 @@ public class QueryOptionController {
     }
 
     @GetMapping("/clearqueryoptions")
-    public void clearQueryOptions()
+    public void clearQueryOptions(String userKey)
     {
+        if (serviceUserRepository.ısThereServiceUser(userKey, (byte) 2, ServiceUser.adminKey))
+        {
+            return;
+        }
         queryOptionRepository.deleteAll();
     }
 }
