+++ src/main/java/com/example/weatherforecastservice/controller/SourceServiceController.java
@@ -1,9 +1,13 @@
 package com.example.weatherforecastservice.controller;
 
+import com.example.weatherforecastservice.model.ServiceUser;
 import com.example.weatherforecastservice.model.SourceService;
+import com.example.weatherforecastservice.repository.ServiceUserRepository;
 import com.example.weatherforecastservice.repository.SourceServiceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
@@ -12,30 +16,36 @@ import java.util.List;
 public class SourceServiceController {
     @Autowired
     private SourceServiceRepository sourceServiceRepository;
+    @Autowired
+    private ServiceUserRepository serviceUserRepository;
 
-    @GetMapping("/addsourceservice")
-    public SourceService addCityService(String apiKey, Long dailyLimit, String url, String  description)
+    @PostMapping("/addsourceservice")
+    public SourceService addCityService(String userKey, @RequestBody SourceService sourceService)
     {
-        SourceService sourceService = new SourceService();
-        sourceService.setDescription(description);
-        sourceService.setServiceURL(url);
-        sourceService.setDailyLimit(dailyLimit);
-        sourceService.setAPIKey(apiKey);
-
-        sourceServiceRepository.save(sourceService);
-
-        return sourceService;
+        if (serviceUserRepository.ısThereServiceUser(userKey, (byte) 1, ServiceUser.adminKey))
+        {
+            return null;
+        }
+        return sourceServiceRepository.save(sourceService);
     }
 
     @GetMapping("/deletesourceservice")
-    public void deleteSourceService(Long id)
+    public void deleteSourceService(String userKey, Long id)
     {
+        if (serviceUserRepository.ısThereServiceUser(userKey, (byte) 1, ServiceUser.adminKey))
+        {
+            return;
+        }
         sourceServiceRepository.deleteById(id);
     }
 
     @GetMapping("/listsourceservices")
-    public List<SourceService> listSourceServices ()
+    public List<SourceService> listSourceServices (String userKey)
     {
+        if (serviceUserRepository.ısThereServiceUser(userKey, (byte) 1, ServiceUser.adminKey))
+        {
+            return null;
+        }
         return sourceServiceRepository.findAll();
     }
